Don't combine default/persistent fields with developer provided ones
We don't want all fields to be persistent, meaning, to be defined as
if they were defined for all subcommands. Connectors will define
required fields, and making these persistent means that they will be
required in every subcommand, even in those that they make no sense
like `help` or `completions`.

So, anyway, this also implied some changes in design, like not mixing
fields defined by the developer of the connector. The reason for this
is that right now those fields that are persistent are the default
ones brought by the SDK, and Cobra has problems making relationships
between fields if some of them are defined on the subcommand while
others are not. Besides, only the main command should get defined
those fields that are persistent.