Return partial body. (#239)
If there's an error reading the response but we got some data, set it so the caller can read it.

This is needed to work around a bug in Percipio's API. If the response is large enough, it returns an incorrect content length and closes the connection unexpectedly. The response contains the full JSON payload... most of the time.