Limit(ed) interfaces for connectorbuilders...
The existing, exported CB interfaces embed resource syncers or sometimes connectorbuilder itself.
This doesn't work with a CBV2, and we strictly never use the embeded interfaces.
Additionally, we sometimes redeclare a "limited" interface inside CB methods.

This commit moves the internal maps to using a limited interface that does not embed syncers/cb
while preserving the older interfaces in case they are randomly used in connectors.