Grant expansion: Remove expandable annotation after processing expandable grant. (#601)
When compacting syncs, we spend a lot of time expanding grants that have already been expanded. So now we remove the expandable annotation after putting the grant in the graph.
Also improved the grant expand test to validate that we get the correct number of grants and that expandable annotations are removed.